#19 Added support of kwargs for remote engines

+++ engines/mixql-engine-sqlite/src/main/scala/org/mixql/engine/sqlite/SqliteSimpleProc.scala
@@ -11,17 +11,17 @@ object SqliteSimpleProc {
     }
 
   val simple_func_params =
-    new ((String, Int) => String) {
+    new ((String, Long) => String) {
 
-      override def apply(a: String, b: Int): String = {
+      override def apply(a: String, b: Long): String = {
         s"SUCCESS:$a:${b.toString}"
       }
     }
 
   val simple_func_context_params =
-    new ((SQLightJDBC, String, Int) => String) {
+    new ((SQLightJDBC, String, Long) => String) {
 
-      override def apply(ctx: SQLightJDBC, a: String, b: Int): String = {
+      override def apply(ctx: SQLightJDBC, a: String, b: Long): String = {
         s"SUCCESS:${ctx.getClass}:$a:${b.toString}"
       }
     }
