+++ engines/mixql-engine-stub/src/main/scala/StubSimpleProc.scala
@@ -16,17 +16,17 @@ object StubSimpleProc {
     }
 
   val simple_func_params =
-    new ((String, Int) => String) {
+    new ((String, Long) => String) {
 
-      override def apply(a: String, b: Int): String = {
+      override def apply(a: String, b: Long): String = {
         s"SUCCESS:$a:${b.toString}"
       }
     }
 
   val simple_func_context_params =
-    new ((StubContext, String, Int) => String) {
+    new ((StubContext, String, Long) => String) {
 
-      override def apply(ctx: StubContext, a: String, b: Int): String = {
+      override def apply(ctx: StubContext, a: String, b: Long): String = {
         s"SUCCESS:${ctx.name}:$a:${b.toString}"
       }
     }
@@ -114,9 +114,9 @@ object StubSimpleProc {
 
   // closure
   val execute_stub_func_using_platform_in_stub_func =
-    new ((PlatformContext, String, Int) => String) {
+    new ((PlatformContext, String, Long) => String) {
 
-      override def apply(ctx: PlatformContext, a: String, b: Int): String = {
+      override def apply(ctx: PlatformContext, a: String, b: Long): String = {
         val funcArgs = List(new MString(a), new MInt(b))
         s"CLOSURE:${ctx.invokeFunction("stub_simple_proc_context_params", funcArgs).asInstanceOf[MString]}"
       }
